Refactor User test models
