+++ cacheio-core/src/test/java/com/mobilejazz/cacheio/model/UserTestModelNotSerializable.java
@@ -1,26 +1,4 @@
 package com.mobilejazz.cacheio.model;
 
-public class UserTestModelNotSerializable {
-
-  private int id;
-  private String name;
-
-  public UserTestModelNotSerializable() {
-  }
-
-  public int getId() {
-    return id;
-  }
-
-  public void setId(int id) {
-    this.id = id;
-  }
-
-  public String getName() {
-    return name;
-  }
-
-  public void setName(String name) {
-    this.name = name;
-  }
+public class UserTestModelNotSerializable extends UserTestModel {
 }
